updated total-supply logic
